File renames + update components
