+++ front/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Link, Router } from "react-router-dom";
 import logo from './assets/images/logo.svg';
-import Home from './pages/Home';
+import Home from './pages/Homepage';
 import Options from './pages/Options';
 import './assets/css/App.css';
 import { io, Socket } from 'socket.io-client';
